+++ docs/versionhistory.rst
@@ -6,6 +6,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 **UNRELEASED**
 
 - Changed the ``repr()`` of ``ExceptionGroup`` to match trio's ``MultiError``
+- Fixed ``TLSStream`` not relaying the extra attributes from its transport stream
+- Fixed incorrect type annotations of some ``TLSAttribute`` members
 
 **2.0.0rc1**